Habilitar borrar Tokens por cada solicitud que se realice a la API

+++ app/Providers/AuthServiceProvider.php
@@ -35,9 +35,9 @@ class AuthServiceProvider extends ServiceProvider
             return 'App\\Policies\\Api\\v1\\'. Pluralizer::singular(class_basename($model)).'Policy';
         });
 
-        /*if (Schema::hasTable('personal_access_tokens')) {
+        if (Schema::hasTable('personal_access_tokens')) {
             AuthService::RemoveExpiredTokensAction();
-        }*/
+        }
 
     }
 }
